+++ config/importmap.rb
@@ -4,3 +4,4 @@ pin "turbo_overlay/stack_controller", to: "turbo_overlay/stack_controller.js", p
 pin "turbo_overlay/overlay_controller", to: "turbo_overlay/overlay_controller.js", preload: true
 pin "turbo_overlay/hint", to: "turbo_overlay/hint.js", preload: true
 pin "turbo_overlay/popover_position", to: "turbo_overlay/popover_position.js", preload: true
+pin "turbo_overlay/history", to: "turbo_overlay/history.js", preload: true